Turn showHide into an arrow function
