+++ index.html
@@ -6,7 +6,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>A complete website with navigation</title>
   <link rel="stylesheet" href="index.css">
-  <!-- <script type="module" src="index.js"></script> -->
 </head>
 <body>
 
@@ -54,8 +53,8 @@
     </div>
   </footer>
 
-  <!-- <script src="./modules/display.js"></script> -->
   <script type="module" src="index.js"></script>
+  <script type="module" src="./modules/content.js"></script>
 
 </body>
 </html>
