+++ modules/content.js
@@ -1,4 +1,4 @@
-export default function showHide(flag)  {
+const showHide = flag => {
   document.getElementById('page1').style.display = 'none';
   document.getElementById('page2').style.display = 'none';
   document.getElementById('page3').style.display = 'none';
@@ -13,3 +13,19 @@ export default function showHide(flag)  {
       document.getElementById('page3').style.display = 'block';
   }
 }
+
+// Adding the correct section
+/* eslint-disable no-unused-vars */
+
+const list = document.querySelector('#list');
+list.addEventListener(('click'), () => { console.log('hit list'); showHide(1) });
+
+const add = document.querySelector('#add-new');
+add.addEventListener(('click'), () => { showHide(2) });
+
+const cont = document.querySelector('#info-contact');
+cont.addEventListener(('click'), () => { showHide(3) });
+
+document.addEventListener('DOMContentLoaded', showHide(1));
+
+export default showHide;
